+++ src/config/server.js
@@ -2,6 +2,7 @@ import express from 'express';
 
 import indexRoute from '../routes/index.routes.js';
 import comunasRoute from '../routes/comunas.routes.js';
+import provinciasRoute from '../routes/provincias.routes.js';
 //const express = require('express');
 //require('dotenv').config();
 
@@ -31,6 +32,7 @@ class Server{
     routes=()=>{
         this.express.use('/', indexRoute);
         this.express.use('/comunas',comunasRoute);
+        this.express.use('/provincias',provinciasRoute);
     }
 
     listen=()=>{
